Use separate models for offer.
