+++ huntflow_api_client/models/request/applicant_offers.py
