+++ tests/test_entities/test_account_offers.py
@@ -3,20 +3,16 @@ from typing import Any, Dict
 from pytest_httpx import HTTPXMock
 
 from huntflow_api_client import HuntflowAPI
-from huntflow_api_client.entities.offers import Offer
-from huntflow_api_client.models.request.offers import ApplicantOfferUpdate
-from huntflow_api_client.models.response.offers import (
+from huntflow_api_client.entities.account_offers import AccountOffer
+from huntflow_api_client.models.response.account_offers import (
     AccountOfferResponse,
     AccountOffersListResponse,
-    ApplicantVacancyOfferResponse,
 )
 from huntflow_api_client.tokens.proxy import HuntflowTokenProxy
 from tests.api import BASE_URL
 
 ACCOUNT_ID = 1
 OFFER_ID = 2
-APPLICANT_ID = 3
-VACANCY_FRAME_ID = 4
 
 GET_ORG_OFFERS_RESPONSE: Dict[str, Any] = {
     "items": [
@@ -116,19 +112,6 @@ GET_ORG_OFFERS_WITH_SCHEMA_RESPONSE: Dict[str, Any] = {
         },
     },
 }
-OFFER_RESPONSE: Dict[str, Any] = {
-    "id": 1,
-    "account_applicant_offer": 1,
-    "created": "2020-01-01T00:00:00+03:00",
-    "values": {
-        "whom_name": "Doe John Hovard",
-        "position_name": "CTO",
-        "whom_date": "24.11.2021",
-        "offer_text": "<p>Welcome to our company!</p>",
-    },
-}
-
-GET_PDF_RESPONSE = bytes("pdf", "UTF-8")
 
 
 async def test_list(
@@ -140,7 +123,7 @@ async def test_list(
         json=GET_ORG_OFFERS_RESPONSE,
     )
     api_client = HuntflowAPI(BASE_URL, token_proxy=token_proxy)
-    offers = Offer(api_client)
+    offers = AccountOffer(api_client)
 
     response = await offers.list(account_id=ACCOUNT_ID)
     assert response == AccountOffersListResponse(**GET_ORG_OFFERS_RESPONSE)
@@ -155,75 +138,7 @@ async def test_get(
         json=GET_ORG_OFFERS_WITH_SCHEMA_RESPONSE,
     )
     api_client = HuntflowAPI(BASE_URL, token_proxy=token_proxy)
-    offers = Offer(api_client)
+    offers = AccountOffer(api_client)
 
     response = await offers.get(account_id=ACCOUNT_ID, offer_id=OFFER_ID)
     assert response == AccountOfferResponse(**GET_ORG_OFFERS_WITH_SCHEMA_RESPONSE)
-
-
-async def test_update(
-    httpx_mock: HTTPXMock,
-    token_proxy: HuntflowTokenProxy,
-) -> None:
-    httpx_mock.add_response(
-        url=f"{BASE_URL}/accounts/{ACCOUNT_ID}/applicants/{APPLICANT_ID}/offers/{OFFER_ID}",
-        json=OFFER_RESPONSE,
-    )
-    api_client = HuntflowAPI(BASE_URL, token_proxy=token_proxy)
-    offers = Offer(api_client)
-    data = ApplicantOfferUpdate(
-        account_applicant_offer=1,
-        values={
-            "position_name": "CTO",
-            "offer_text": "<p>Welcome to our company!</p>",
-            "whom_date": "2021-11-24",
-            "whom_name": "Doe John Hovard",
-        },
-    )
-
-    response = await offers.update(
-        account_id=ACCOUNT_ID,
-        applicant_id=APPLICANT_ID,
-        offer_id=OFFER_ID,
-        data=data,
-    )
-    assert response == ApplicantVacancyOfferResponse(**OFFER_RESPONSE)
-
-
-async def test_get_applicant_on_vacancy_offer(
-    httpx_mock: HTTPXMock,
-    token_proxy: HuntflowTokenProxy,
-) -> None:
-    httpx_mock.add_response(
-        url=f"{BASE_URL}/accounts/{ACCOUNT_ID}/applicants/{APPLICANT_ID}"
-        f"/vacancy_frames/{VACANCY_FRAME_ID}/offer",
-        json=OFFER_RESPONSE,
-    )
-    api_client = HuntflowAPI(BASE_URL, token_proxy=token_proxy)
-    offers = Offer(api_client)
-
-    response = await offers.get_applicant_on_vacancy_offer(
-        account_id=ACCOUNT_ID,
-        applicant_id=APPLICANT_ID,
-        vacancy_frame_id=VACANCY_FRAME_ID,
-    )
-    assert response == ApplicantVacancyOfferResponse(**OFFER_RESPONSE)
-
-
-async def test_get_offer_pdf(
-    httpx_mock: HTTPXMock,
-    token_proxy: HuntflowTokenProxy,
-) -> None:
-    httpx_mock.add_response(
-        url=f"{BASE_URL}/accounts/{ACCOUNT_ID}/applicants/{APPLICANT_ID}/offers/{OFFER_ID}/pdf",
-        content=GET_PDF_RESPONSE,
-    )
-    api_client = HuntflowAPI(BASE_URL, token_proxy=token_proxy)
-    offers = Offer(api_client)
-
-    response = await offers.get_offer_pdf(
-        account_id=ACCOUNT_ID,
-        applicant_id=APPLICANT_ID,
-        offer_id=OFFER_ID,
-    )
-    assert response == GET_PDF_RESPONSE
